requirements.txt
# Expiry Tracker - Dependencies for Render.com
# Web framework
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-migrate>=4.0.0
gunicorn>=21.0.0

# Database connectivity
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# MongoDB connectivity (for migration)
pymongo>=4.5.0

# Environment and configuration
python-dotenv>=1.0.0

# Data processing
python-dateutil>=2.8.0

# Image processing for camera feature
Pillow>=10.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0